+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Custom labels for each change type
 ### Updated
 - Improve and speed up the way diff file contents are shown
+- Renamed some settings
 - Some libraries
 ### Fixed
 - Not handling correctly file paths with spaces in their names